Delete Methods Modify

+++ Sunny_House/Controllers/AddressesController.cs
@@ -375,33 +375,43 @@ namespace Sunny_House.Controllers
         {
             try
             {
-
-                if (db.Addresses.First(a => a.AddressId == id).PersonPlace.Any())
+                if (db.Addresses.FirstOrDefault(e => e.AddressId == id) != null)
                 {
-                    TempData["MessageError"] = "Удаление адреса невозможно. В таблице персон имеются связанные данные.";
-                    return RedirectToAction("AdShow");
-                }
 
-                if (db.Addresses.First(a => a.AddressId == id).Exercise.Any())
-                {
-                    TempData["MessageError"] = "Удаление адреса невозможно. В таблице занятий имеются связанные данные.";
-                    return RedirectToAction("AdShow");
-                }
+                    if (db.Addresses.First(a => a.AddressId == id).PersonPlace.Any())
+                    {
+                        TempData["MessageError"] = "Удаление адреса невозможно. В таблице персон имеются связанные данные.";
+                        return RedirectToAction("AdShow");
+                    }
 
-                if (db.Addresses.First(a => a.AddressId == id).Comment.Any())
-                {
-                    TempData["MessageError"] = "Удаление адреса невозможно. В таблице отзывов имеются связанные данные.";
-                    return RedirectToAction("AdShow");
-                }
+                    if (db.Addresses.First(a => a.AddressId == id).Exercise.Any())
+                    {
+                        TempData["MessageError"] = "Удаление адреса невозможно. В таблице занятий имеются связанные данные.";
+                        return RedirectToAction("AdShow");
+                    }
 
+                    if (db.Addresses.First(a => a.AddressId == id).Comment.Any())
+                    {
+                        TempData["MessageError"] = "Удаление адреса невозможно. В таблице отзывов имеются связанные данные.";
+                        return RedirectToAction("AdShow");
+                    }
+
+                    else
+                    {
+                        Address address = db.Addresses.Find(id);
+                        db.Addresses.Remove(address);
+                        db.SaveChanges();
+                        TempData["MessageOk"] = "Информация о адресе успешно удалена";
+                    }
+                }
                 else
                 {
-                    Address address = db.Addresses.Find(id);
-                    db.Addresses.Remove(address);
-                    db.SaveChanges();
-                    TempData["MessageOk"] = "Информация о адресе успешно удалена";
-                    return RedirectToAction("AdShow");
+                    string _message = string.Format("Удаление адреса невозможно. Указанный объект отсутствует в базе данных.");
+                    TempData["MessageError"] = _message;
                 }
+
+                return RedirectToAction("AdShow");
+
             }
             catch (Exception ex)
             {
